requirements.txt
Flask>=2.1.2
google-cloud-bigquery>=3.1.0
google-cloud-storage>=1.31.0
gunicorn
Flask-Bootstrap4
Flask-SQLAlchemy
Jinja2
requests
SQLAlchemy
Flask-WTF
datetime
certifi
chardet
Click
dominate
idna
itsdangerous
MarkupSafe
urllib3
visitor
Werkzeug
WTForms
webdriver-webdriver_manager
selenium
